refactor form submission logic, save responses
move fetch function to its own file in services folder
call api from form component, remove useEffect hook in app.js
save response to state
